src/requirements.txt
Flask~=2.0.2
evolved5g~=0.8.8
jsonpickle~=2.0.0
Pillow~=9.0.0
python-dotenv~=0.21.0
python-statemachine~=1.0.3
